Create RHOK Code for Boston NASA Dark Skies
Dark Skies project
The challenge was to crowdsource locations for optimal viewing of celestial events. This was approached from two complementary angles. One team set out to use population density as a means of producing a map where ambient light pollution would be limited. A second team sought to create an app using Google's Open Source starmap with an overlay having a simple transparency slider. In this way, users align their device with a patch of night sky, adjust the slider until the screen matches their view and touch a "send" function. The position, elevation, angle and transparency level are then sent to a central map environment for community viewing and interaction. Based upon each parameter, a fairly accurate assessment of ambient light levels can be calculated for each location. The product from both teams become layers that could be added to by individual reports and contribute to similar projects globally. 
We accomplished loading the starmap on an Android device, providing an overlay with a transparency controller and site mockups. Inquiries: om@standbytaskforce.com